+++ data/beta/archive/Lodestar_Guide-adf01e92.lua
@@ -1,0 +1,132 @@
+
+LodestarScans = {
+["build"] = "69893",
+["scan"] = {
+},
+["trails"] = {
+[2521] = {
+["p"] = 0,
+["rows"] = {
+[83] = "07d00",
+},
+["n"] = 1,
+["l"] = 0,
+["nx"] = 278,
+["ny"] = 185,
+},
+},
+["v"] = 1,
+}
+LodestarHarvest = {
+["objects"] = {
+},
+["levels"] = {
+[6] = 3600,
+},
+["meta"] = {
+["contributors"] = {
+["Venz Dru-ClassicBetaPvP2"] = {
+["race"] = "Skyborne",
+["last"] = 1789800053,
+["faction"] = "Horde",
+["sessions"] = 1,
+["first"] = 1789800053,
+["level"] = 6,
+["class"] = "DRUID",
+},
+},
+["build"] = "69893",
+["v"] = 1,
+},
+["npcs"] = {
+[254360] = {
+["map"] = 2521,
+["ctype"] = "Humanoid",
+["samples"] = {
+{
+2521,
+45.2,
+45.3,
+"Zephras Isle",
+"Shen'dar Village",
+},
+},
+["near"] = 1,
+["y"] = 45.3,
+["x"] = 45.2,
+["name"] = "Belandiel Farflight",
+["seen"] = 1,
+["react"] = 5,
+["maxL"] = 10,
+["minL"] = 10,
+},
+[254100] = {
+["map"] = 2521,
+["ctype"] = "Humanoid",
+["samples"] = {
+{
+2521,
+45.2,
+45.3,
+"Zephras Isle",
+"Shen'dar Village",
+},
+},
+["near"] = 1,
+["y"] = 45.3,
+["x"] = 45.2,
+["name"] = "Zephras Citizen",
+["seen"] = 1,
+["react"] = 5,
+["maxL"] = 10,
+["minL"] = 10,
+},
+},
+["taxi"] = {
+},
+["build"] = "69893",
+["quests"] = {
+[92514] = {
+["classes"] = {
+["DRUID"] = true,
+},
+["races"] = {
+["Skyborne"] = true,
+},
+["lvl"] = 6,
+["t"] = "Welcome to Shen'dar Village",
+["o"] = {
+{
+["text"] = "Speak with Illaya Amberwind",
+["type"] = "monster",
+["n"] = 1,
+},
+{
+["text"] = "Speak with the Innkeeper",
+["type"] = "monster",
+["n"] = 1,
+},
+},
+},
+[97968] = {
+["classes"] = {
+["DRUID"] = true,
+},
+["races"] = {
+["Skyborne"] = true,
+},
+["lvl"] = 6,
+["t"] = "Camping 101: Herbalism",
+["o"] = {
+{
+["text"] = "Raise your herbalism skill to 20",
+["type"] = "object",
+["n"] = 1,
+},
+},
+},
+},
+["v"] = 1,
+}
+LodestarScanDB = nil
+LodestarShareDB = nil